LayoutTests/fast/dom/HTMLKeygenElement/keygen-expected.txt


This tests the keygen element.

PASS keygen.options is undefined
PASS shadow.tagName is "SELECT"
PASS layoutTestController.shadowRoot(shadow) is null
PASS successfullyParsed is true

TEST COMPLETE
